test: converts ava config to file based
Run commands were using glob patterns, e.g. "test/*.js", which are
expanded by shell on unix systems and thus made the tests run. However,
this isn't the correct way to use globs with ava, and the correct way is
to use the config option "files" to specify glob patterns. If we use CLI
based glob patterns they are limited to (1) shell expansion limits and
(2) do not work on Windows. Because this project has two types of tests
for js and ts files, it was not possible to use inline test config
inside package.json, and we have to rely on file-based configs and
specify them via --config option.